agents/requirements.txt
# Google Generative AI (Gemini)
google-generativeai>=0.8.0

# Google ADK for Agent Development (optional, for advanced orchestration)
# google-adk>=1.0.0

# API framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# PDF/Document handling for CV
pypdf2>=3.0.0
python-docx>=1.1.0
reportlab>=4.0.0

# Job APIs
httpx>=0.26.0
